% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ct.modelling-package.R
\docType{package}
\name{ct.modelling-package}
\alias{ct.modelling}
\alias{ct.modelling-package}
\title{ct.modelling: Viral Load Curve Estimation}
\description{
Viral load curve estimation
}
\author{
\strong{Maintainer}: Tim Russell \email{timothy.russell@lshtm.ac.uk}

}
\keyword{internal}
